requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
websockets==13.0.1
twilio==9.3.3
streamlit==1.41.1
pydantic==2.9.2
requests==2.32.3
